+++ The Soccer Challager/main-scene.js
@@ -10,7 +10,7 @@ import {Transforms_Sandbox} from "./examples/transforms-sandbox.js"
 import {Shadow_Demo} from "./examples/shadow-demo.js"
 import {Control_Demo} from "./examples/control-demo.js"
 import {Mouse_Demo} from "./examples/mouse-demo.js"
-import {Main} from "./humanfigure.js";
+import {Main} from "./SoccerChallenge.js";
 // Pull these names into this module's scope for convenience:
 const {
     Vector, Vector3, vec, vec3, vec4, color, Matrix, Mat4, Light, Shape, Material, Shader, Texture, Scene,
